requirements.txt
streamlit==1.50.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
pandas==2.2.3
requests==2.32.5
requests-oauthlib==2.0.0
mysql-connector-python==9.4.0
typing_extensions==4.12.2
pydantic==2.10.6
pydantic_core==2.27.2
uvicorn==0.37.0